refactor: splitting the generator
This is mostly minor, but since we'll be extending this (to generate Package Definitions,
API Version Definitions and Service Definitions) this seems worthwhile to do first
